requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenvfastapi[all]
